# Conflicts:
#	app/src/main/java/com/pantaubersama/app/ui/profile/verifikasi/step3/Step3VerifikasiActivity.kt
#	app/src/main/java/com/pantaubersama/app/ui/profile/verifikasi/step5/Step5VerifikasiActivity.kt
#	app/src/main/java/com/pantaubersama/app/ui/profile/verifikasi/step7/Step7VerifikasiActivity.kt
#	app/src/main/java/com/pantaubersama/app/utils/ImageUtil.kt

+++ app/src/main/java/com/pantaubersama/app/base/BaseRecyclerAdapter.kt
@@ -109,7 +109,7 @@ abstract class BaseRecyclerAdapter : RecyclerView.Adapter<RecyclerView.ViewHolde
     fun setLoaded() {
         if (isLoadingMore) {
             isLoadingMore = false
-            deleteItem(itemCount - 1)
+            if (data[itemCount - 1] is LoadingModel) deleteItem(itemCount - 1)
         }
     }
